pyproject.toml
[project]
name = "speakup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.134.0",
    "faster-whisper>=1.2.1",
    "keyboard>=0.13.5",
    "llama-cpp-python>=0.3.16",
    "loguru>=0.7.3",
    "mss>=10.1.0",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pyperclip>=1.11.0",
    "pystray>=0.19.5",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "pyttsx3>=2.99",
    "pywin32>=311",
    "pywinauto>=0.6.9",
    "requests>=2.32.5",
    "sounddevice>=0.5.5",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
